+++ src/Request/Internal.php
@@ -313,7 +313,8 @@ class Internal extends RequestCollection
                 $request
                     ->addPost('client_shared_at', (string) time())
                     ->addPost('source_type', '3')
-                    ->addPost('configure_mode', Constants::SHARE_TYPE['REEL_SHARE']) // 2 - REEL_SHARE
+                    ->addPost('configure_mode', 1)
+                    //->addPost('configure_mode', Constants::SHARE_TYPE['REEL_SHARE']) // 2 - REEL_SHARE
                     ->addPost('client_timestamp', (string) (time() - mt_rand(3, 10)))
                     ->addPost('upload_id', $uploadId);
 
@@ -396,7 +397,7 @@ class Internal extends RequestCollection
                     ->addPost('thread_ids', $internalMetadata->getDirectThreads())
                     ->addPost('client_shared_at', (string) time())
                     ->addPost('source_type', '3')
-                    ->addPost('configure_mode', Constants::SHARE_TYPE['DIRECT_STORY_SHARE']) // 3 - DIRECT_STORY_SHARE
+                    ->addPost('configure_mode', Constants::SHARE_TYPE['REEL_SHARE']) // 2 - REEL_SHARE (stories)
                     ->addPost('client_timestamp', (string) (time() - mt_rand(3, 10)))
                     ->addPost('upload_id', $uploadId);
 
@@ -786,7 +787,8 @@ class Internal extends RequestCollection
                 }
 
                 $request
-                    ->addPost('configure_mode', Constants::SHARE_TYPE['REEL_SHARE']) // 2 - REEL_SHARE
+                    ->addPost('configure_mode', 1)
+                    //->addPost('configure_mode', Constants::SHARE_TYPE['REEL_SHARE']) // 2 - REEL_SHARE
                     ->addPost('story_media_creation_date', time() - mt_rand(10, 20))
                     ->addPost('client_shared_at', time() - mt_rand(3, 10))
                     ->addPost('client_timestamp', time());
@@ -876,7 +878,7 @@ class Internal extends RequestCollection
                 break;
             case Constants::FEED_DIRECT_STORY:
                 $request
-                    ->addPost('configure_mode', Constants::SHARE_TYPE['DIRECT_STORY_SHARE']) // 3 - DIRECT_STORY_SHARE
+                    ->addPost('configure_mode', Constants::SHARE_TYPE['REEL_SHARE']) // 2 - REEL_SHARE
                     ->addPost('recipient_users', $internalMetadata->getDirectUsers())
                     ->addPost('thread_ids', $internalMetadata->getDirectThreads())
                     ->addPost('story_media_creation_date', time() - mt_rand(10, 20))
@@ -899,8 +901,8 @@ class Internal extends RequestCollection
                 }
                 $request
                     ->addPost('title', $title)
-                    ->addPost('caption', $captionText)
-                    ->addPost('configure_mode', Constants::SHARE_TYPE['IGTV']); // 8 - IGTV
+                    ->addPost('caption', $captionText);
+                    //->addPost('configure_mode', Constants::SHARE_TYPE['IGTV']); // 8 - IGTV
                 break;
         }